Moved a few classes to their own module and removed the host provider
