+++ arteria/logging.py
@@ -1,0 +1,6 @@
+class Logger:
+    def debug(self, msg): print msg 
+    def info(self, msg): print msg
+    def warn(self, msg): print msg
+    def error(self, msg): print msg
+
